feat: refactor ImageUploader component atomized and tests adapted
