+++ src/components/shared/ui/image-uploader/parts/UploaderLabel.tsx
@@ -1,0 +1,7 @@
+export function UploaderLabel({ label }: { label: string }) {
+  return (
+    <label className='text-[10px] uppercase tracking-[0.2em] font-bold text-neutral-500'>
+      {label}
+    </label>
+  )
+}
