+++ tests/features/storage/hooks/useImageUploader.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import { renderHook, act } from '@testing-library/react'
-import { useImageUploader } from '@/components/shared/ui/hooks/useImageUploader'
+import { useImageUploader } from '@/components/shared/ui/image-uploader/hooks/useImageUploader'
 import { StorageService } from '@/services/storage-service'
 import { useAuth } from '@clerk/nextjs'
 import { sileo } from 'sileo'
